refactor main/index.ts
